代码优化： 重构 join node

+++ template/install_docker.sh
@@ -1,9 +1,17 @@
 #!/usr/bin/env bash
+# shellcheck disable=SC2006
+export VERSION=`docker version  --format="{{.Server.Version}}"`
+if [[ $VERSION == "20.10.24" ]];then
+  echo "docker version 20.10.24 already install"
+  exit 0
+fi
+
+set -e
 echo "clean env"
 yum remove -y docker docker-common container-selinux docker-selinux docker-engine
 rm -rf /var/lib/docker
 
-echo "install docker 19.04.14"
+echo "install docker 20.10.24"
 yum install -y yum-utils
 
 yum-config-manager \
@@ -12,15 +20,15 @@ yum-config-manager \
 
 yum clean packages
 #查看docker-ce版本并且安装
-yum list docker-ce --showduplicates | sort -r  
-yum install -y docker-ce-cli-19.03.14 docker-ce-19.03.14 containerd.io
+yum list docker-ce --showduplicates | sort -r
+yum install -y docker-ce-cli-20.10.24 docker-ce-20.10.24 containerd.io docker-compose-plugin docker-ce-rootless-extras-20.10.24
 
 
 echo "config docker daemon"
 mkdir -p /etc/docker
 cat > /etc/docker/daemon.json <<EOF
 {
-  "data-root": "/data/docker",
+  "data-root": "/var/lib/docker",
   "storage-driver": "overlay2",
   "exec-opts": [
     "native.cgroupdriver=systemd",
